remove data provider to merge in with census embedded provider
